Updated Links to Short URLs And README
The commands, and post of the day, now post the short link, or the
reddit redirect link instead of the full post name. Lots of channels
have bots that parse page titles, so this will allow them to do their
thing without blowing up the chat.

Also updated the README to be more accurate to what is available, and
easier to read.