Add New Param for Top Posts And Update Packages
Added `timeframe` parameter to the `!top` command, which is optional. It
is case insensitive, and handles many different cases for how to note
the timeframe to return posts for. Updated the README to reflect this
change so that command will be known.

Updated `better-sqlite3` and `phin` to current versions to keep this
project up to date on its dependencies. Currently, this project is not
running through any semantic versioning of its own though.

Also, snuck in a small change to how the top post of the day is posted.
Now sent in a single message since the short link is being used.